refactor(solver): split query cache statistics module
