+++ crates/tsz-solver/src/caches/mod.rs
@@ -2,5 +2,6 @@ pub(crate) mod db;
 pub(crate) mod instantiation_cache;
 pub(crate) mod query_cache;
 mod query_cache_evaluation;
+pub(crate) mod query_cache_statistics;
 pub(crate) mod query_trace;
 pub(crate) mod subtype_reduction_cache;
